+++ ttmp/2026/04/25/PYXIS-CSSVD-JS-LIB--implement-pyxis-css-visual-diff-javascript-userland-library/scripts/lib/index.js
@@ -6,5 +6,8 @@ module.exports = {
   policies: require('./policies.js'),
   markdown: require('./markdown.js'),
   results: require('./results.js'),
+  styles: require('./styles.js'),
+  inspect: require('./inspect.js'),
 }
+